Refactor test_models
